==> 00_Data-Collection.ipynb <==
import pandas as pd
import GetOldTweets3 as got
# https://github.com/Mottl/GetOldTweets3

# function to get twitter info and organize in dataframe
def get_twitter_info():
    tweet_df["id"] = tweet_df["got_criteria"].apply(lambda x: x.id)
    tweet_df["tweet_text"] = tweet_df["got_criteria"].apply(lambda x: x.text)
    tweet_df["timestamp"] = tweet_df["got_criteria"].apply(lambda x: x.date)
    tweet_df["hashtags"] = tweet_df["got_criteria"].apply(lambda x: x.hashtags)
    tweet_df["username"] = tweet_df["got_criteria"].apply(lambda x: x.username)
    tweet_df["mentions"] = tweet_df["got_criteria"].apply(lambda x: x.mentions)
    tweet_df["rewtweets"] = tweet_df["got_criteria"].apply(lambda x: x.retweets)
    tweet_df["replies"] = tweet_df["got_criteria"].apply(lambda x: x.replies)
    tweet_df["link"] = tweet_df["got_criteria"].apply(lambda x: x.permalink)
# https://medium.com/@robbiegeoghegan/download-twitter-data-with-10-lines-of-code-42eb2ba1ab0f

*NOTE:* the code below creates a dateframe based on the features in the function above. The dataframes are location-based. We received error messages when saving all cities into the `locations` variable. As manual workaround, we updated each city in the `locations` variable manually.

The cities included are as follows: `Butte County`, `Paradise`, `Chico`, `Durham`, `Oroville`, `Magalia`. This is pulled from the location tag of a given tweet.

#specify location
locations = ["Paradise, CA"]

#get old tweets
##date range pulled from for fire: 11/8/2018 - 11/26/2018
##date range pulled from pre fire: 11/1/2018 - 11/7/2018
tweetCriteria_list = []
for location in locations:
    tweetCriteria = got.manager.TweetCriteria().setSince('2018-11-08')\
                                            .setUntil('2018-11-26')\
                                            .setNear(location)\
                                            .setWithin('100mi')
    tweetCriteria_list.append(tweetCriteria)
    
#create twitter info for each city
tweet_dict = {}
for criteria, location in zip(tweetCriteria_list, locations):
    tweets = got.manager.TweetManager.getTweets(criteria)
    tweet_dict[location] = tweets
    
#create df
tweet_df = pd.DataFrame(dict([ (k,pd.Series(v)) for k,v in tweet_dict.items() ]))
tweet_df['tweet_count'] = tweet_df.index
tweet_df = pd.melt(tweet_df, id_vars=["tweet_count"], var_name='City', value_name='got_criteria')
# tweet_df = tweet_df.dropna()

#extract twitter information
get_twitter_info()
tweet_df = tweet_df.drop("got_criteria", 1)

#save copy of df based on city specified in location
paradise = tweet_df.copy()
paradise.shape

#save out raw city csv 
paradise.to_csv('./data/paradise.csv', index=False)

## raw dataframes shape summary

## during fire: 
# paradise (817, 11)
# butte (902, 11)
# chico (811, 11)
# oroville (202, 11)
# magalia (764, 11)
# durham (862, 11)

## pre-fire:
# paradise_pre (145, 11)
# butte_pre (166, 11)
# chico_pre (141, 11)
# oroville_pre (53, 11)
# magalia_pre (144, 11)
# durham_pre (160, 11)


### Create single dataframe for tweets during fire

# read in raw city files
paradise = pd.read_csv('./data/paradise.csv')
butte = pd.read_csv('./data/butte.csv')
chico = pd.read_csv('./data/chico.csv')
magalia = pd.read_csv('./data/magalia.csv')
oroville = pd.read_csv('./data/oroville.csv')

# concatenate raw city files into single df
cities = pd.concat([paradise, butte, chico, durham, magalia, oroville])
cities.shape

# save out raw, concatenated cities file
cities.to_csv('./data/cities_raw.csv', index=False)

# check duplicates
dups = cities[(cities['id'].duplicated() == True)]
dups

# drop duplicates
cities.drop_duplicates(subset ="id", 
                     keep = 'first', inplace = True) 
cities.drop_duplicates(subset ="id", 
                     keep = 'first', inplace = True) 
cities.shape

cities.head()

# save out cleaned cities file 
cities.to_csv('./data/cities_clean.csv', index=False)

### Create single dataframe for tweets 1-week before fire
*repeat same process as above*

paradise_pre = pd.read_csv('./data/paradise_pre.csv')
butte_pre = pd.read_csv('./data/butte_pre.csv')
chico_pre = pd.read_csv('./data/chico_pre.csv')
magalia_pre = pd.read_csv('./data/magalia_pre.csv')
oroville_pre = pd.read_csv('./data/oroville_pre.csv')

pre = pd.concat([paradise_pre, butte_pre, chico_pre, durham_pre, magalia_pre, oroville_pre])
pre.shape

pre.drop_duplicates(subset ="id", 
                     keep = 'first', inplace = True) 
pre.shape

pre.drop_duplicates(subset ="id", 
                     keep = 'first', inplace = True) 
pre.shape

pre.to_csv('./data/pre-fire.csv', index=False)

check_dups = pd.read_csv('./data/pre-fire.csv', index_col=0)

check_dups.shape

check_dups.drop_duplicates(subset ="id", 
                     keep = 'first', inplace = True) 
check_dups.shape

check_dups.to_csv('./data/pre-fire.csv', index=False)

### Create single dataframe for with tweets before and during fire

cities = pd.read_csv('./data/cities_clean.csv')
pre = pd.read_csv('./data/pre-fire.csv')

df = pd.concat([cities, pre])
df.shape

df['timestamp'] = pd.to_datetime(df['timestamp'])

df['day'] = df.timestamp.dt.day
df['hour'] = df.timestamp.dt.hour
df['minute'] = df.timestamp.dt.minute

df.to_csv('./data/stacked_v1.csv')
