==> 01_dictionarybuilding.ipynb <==
import pandas as pd
import numpy as np
import gensim
import requests
from bs4 import BeautifulSoup
import time
import re

# Pulling in tweets that we pulled
df = pd.read_csv('./datasets/stacked_classified.csv')
df.head()

# counts of hashtages (including stripping the '#'
def count_hashtags(df):
    keyterms = {}
    for i in df['hashtags'].dropna():
        i = i.replace("#", '')
        for j in i.split():
            if j not in keyterms.keys():
                keyterms[j] = 1
            else:
                keyterms[j] += 1
    return keyterms

# Created a dictionary of 'root words' related to a wildfire
# Iterate over the hashtags and this list to build a keywords list to score tweets
def build_keywords(hashtags): 
    word_filter = ['fire', 'evac', 'smok', 'burn', 'wild', 'blaz', 'hell', 'department',
              'inferno', 'help']
    keywords = []
    for word in set(hashtags):
        for wf in word_filter:
            if wf in word.lower():
                keywords.append(word.lower())
    return list(set(keywords))

# check each tweet for keyterms and score them
# trying to also account for plural words and capitalizations
# key_score is just the count of the keywords in the tweet
def keyscoring(df, keywords):
    keyscore = []
    df['tweet_text'] = df['tweet_text'].str.lower() #removing all capitals
    for row in df.index:
        keyscore.append(0)
        for word in (df.iloc[row]['tweet_text']).split(): # splitting 1 tweet into words
            word_nohash = re.sub(r'[^\w\s]','', word) #removing all punctuation
            if word_nohash in keywords: # Using try excepts to just make sure nothing breaks with a missing index
                try:
                    keyscore[row]+=1
                except:
                    print(f'fail in row {row}')
            else:
                try:
                    if word_nohash[-1] == 's':
                        if word_nohash[:-1] in keywords:
                            try:
                                keyscore[row]+=1
                            except:
                                print(f'fail in row {row}')
                except:
                    pass
    df['key_score'] = keyscore

# This scrapes each tweet for the from location at the bottom of some tweets
# This is a Twitter generated location if enabled by the user
def get_locs(df):
    locs = []
    df_num_rows = df.shape[0]
    for row in df.index:
        url = df.iloc[row]['link']
        res = requests.get(url)
        soup = BeautifulSoup(res.content, 'lxml')
        spans = soup.find_all('span', {'class' : 'permalink-tweet-geo-text'})
        try:
            location = spans[0].text.replace("from",'').strip()
            if location == 'California, USA':
                locs.append(df.iloc[row]['City'])
            else:
                locs.append(location)
        except:
            print(f'tweet #{row} has no location info')
            locs.append('nolocationfound')
        
        if row % 25 == 0:
            time.sleep(3)
        if row % 75 == 0:
            print(f'Located {row} out of {df_num_rows} tweets.  {row/df_num_rows:.2%}')
    df['from_locations'] = locs

# This just removes the url from the tweet text, this is because URLs can do weird things in NLP
def remove_url(df):
    df['text_nourl'] = [re.sub(r'\w+:\/{2}[\d\w-]+(\.[\d\w-]+)*(?:(?:\/[^\s/]*))*', '', i) for i in df['tweet_text']]

# This runs all the steps to clean an imported tweet dataset
def func_master(df):
    keywords = build_keywords(list(count_hashtags(df)))
    remove_url(df)
    keyscoring(df, keywords)
    get_locs(df)

func_master(df)


df.head() #checking

df['key_score'].value_counts()

df.to_csv('./datasets/stacked_clean_again.csv', index=False) #exporting

### Currently unused

# use hashtag list to build dictionary of keyterms to use combined with that of the gensim model
# Import word vectors into "model."
model = gensim.models.KeyedVectors.load_word2vec_format('../lexvec.enwiki+newscrawl.300d.W.pos.vectors')


# Building a keywords list from related words to a few seed words
keyt = []
keywords = ['fire', 'smoke', 'wildfire', 'campfire', 'forest', 'evacuate', 'hell']
for word in keywords:
    kt = [x[0] for x in model.most_similar(word, topn = 25)]
    keyt+= kt
word_list = list(set(keyt))

# Comparing the keywords to the hashtages to make a list to score tweets
important_words = []
for word in keyterms.keys():
    if word in word_list:
        important_words.append(word)

df[(df['tweet_text'].str.contains('fires')) & (df['key_score'] == 0)]
