==> 02_sentiment.ipynb <==
This notebook is using a [Twitter NLP Toolkit](https://github.com/eschibli/twitter-toolbox) to find the sentiment of the tweets.  It just returns a 1 for positive and 0 for negative

import numpy as np
import pandas as pd
from twitter_nlp_toolkit.tweet_sentiment_classifier import tweet_sentiment_classifier
from twitter_nlp_toolkit.file_fetcher import file_fetcher
from zipfile import ZipFile
import os

Using this library: https://github.com/eschibli/twitter-toolbox

df = pd.read_csv('./datasets/stacked_clean_again.csv')
df.shape

df.head()

#This is adapted from the github linked above
chunk = 1 # Fraction of data to train on - you can reduce for debugging for speed
model_path = '.models' #where the directory is

# loading in the models
classifier = tweet_sentiment_classifier.SentimentAnalyzer()
classifier.load_small_ensemble()

# Testing.  [1, 0, 1, 0] expected
classifier.predict(['I am happy', 'I am sad', 'I am cheerful', 'I am mad'])

# Classifying the tweets
df['sent'] = classifier.predict(df['text_nourl'])

df.to_csv('./datasets/stacked_sentiment_again.csv', index=False) #exporting

# pd.set_option('display.max_colwidth', None)
# pd.set_option('display.max_rows', None)

df2['sent'].value_counts(normalize=True) #checking distribution

# Looking at stats
df2 = df[df['key_score']>0]
df2[['text_nourl', 'sent', 'key_score']]
