==> 03_clustering.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.cluster import KMeans
from sklearn.metrics import silhouette_score
from sklearn.preprocessing import StandardScaler
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer

%matplotlib inline

df = pd.read_csv('./datasets/stacked_sentiment_again.csv')
df.shape

# Trying to find an optimal amount of clusters
def kmeans_check_ks(df, ks):
    vectorizer = TfidfVectorizer(stop_words='english', ngram_range=(1,1))
    X = vectorizer.fit_transform(df['text_nourl'])
    # don't need to scale since Tfidf does that automatically

    scores = []
    for k in range(2, ks+1):
        km = KMeans(n_clusters=k, random_state=2018)
        km.fit(X)
        sil = silhouette_score(X, km.labels_)
        intertia = km.inertia_
        scores.append((k, sil, intertia))

    scores_df = pd.DataFrame(scores, columns=['K', 'SilhouetteScore', 'Inertia'])
    scores_df.sort_values('SilhouetteScore', ascending=False, inplace=True)
    return scores_df

### Basic k-means on all text - We are not using any filters
- Followed steps in [this](https://towardsdatascience.com/applying-machine-learning-to-classify-an-unsupervised-text-document-e7bb6265f52) article

# We need to vectorize first and turn all words into their roots
vectorizer_a = TfidfVectorizer(stop_words='english')
X_all = vectorizer.fit_transform(df['text_nourl'])

# Checking with 3 clusters
true_k = 3
model_all = KMeans(n_clusters=true_k, random_state=2018)
model_all.fit(X)

# Viewing clusters of words
order_centroids = model_all.cluster_centers_.argsort()[:, ::-1]
terms = vectorizer_a.get_feature_names()

for i in range(true_k):
    print(f'Cluster %d:' % i),
    for ind in order_centroids[i, :10]:
        print(f'%s' % terms[ind])

### Basic k-means on fire-related tweets
- follow steps in [this](https://towardsdatascience.com/applying-machine-learning-to-classify-an-unsupervised-text-document-e7bb6265f52) article

# Manually coded each tweet if it was related to the fire.  
# Pulling it into a new dataframe
fire = df[(df['is-fire-related'] == 1)].copy()
fire.shape

# Vectorizing and fitting a Kmeans with 3 clusters
vectorizer_f = TfidfVectorizer(stop_words='english')
X_fire = vectorizer_f.fit_transform(fire['text_nourl'])

true_k = 3
model_3f = KMeans(n_clusters=true_k, random_state=2018)
model_3f.fit(X)

order_centroids = model_3f.cluster_centers_.argsort()[:, ::-1]
terms = vectorizer_f.get_feature_names()

# Viewing clusters
for i in range(true_k):
    print(f'Cluster %d:' % i),
    for ind in order_centroids[i, :10]:
        print(f'%s' % terms[ind])

### 3 clusters
- cluster 0: Traffic and road related
- cluster 1: Emotional
- cluster 2: California

# Fitting with 5 clusters
true_k = 5
model_5f = KMeans(n_clusters=true_k, random_state=2018)
model_5f.fit(X_fire)

order_centroids = model_5f.cluster_centers_.argsort()[:, ::-1]
terms = vectorizer_f.get_feature_names()

# Viewing clusters
for i in range(true_k):
    print(f'Cluster %d:' % i),
    for ind in order_centroids[i, :10]:
        print(f'%s' % terms[ind])

# Fitting with 5 clusters
true_k = 7
model_7f = KMeans(n_clusters=true_k, random_state=2018)
model_7f.fit(X_fire)

order_centroids = model_7f.cluster_centers_.argsort()[:, ::-1]
terms = vectorizer_f.get_feature_names()

# Viewing clusters
for i in range(true_k):
    print(f'Cluster %d:' % i),
    for ind in order_centroids[i, :10]:
        print(f'%s' % terms[ind])

### 5 clusters:
- 0: Traffic 
- 1: Emotional (negative sentiment)
- 2: California
- 3: Traffic
- 4: Informational

# Fitting each tweet to a cluster
fire['clusters'] = fire['text_nourl'].apply(lambda x: int(model_5f.predict(vectorizer_f.transform([x]))[0]))
fire.head()

fire.to_csv('./datasets/clustered_fires.csv', index = False) # exporting

fire[(fire['key_score'] > 0) & (fire['text_nourl'].str.contains('fire'))][['text_nourl', 'key_score', 'clusters']]

fire_scores_2 = kmeans_check_ks(fire, 20)
fire_scores_2

all_clustering = kmeans_check_ks(df, 30)
all_clustering

# Dataframe viewing options for double checking
# pd.set_option('display.max_rows', None)
# pd.set_option('display.max_colwidth', None)


fire.groupby('clusters')['key_score'].mean()

### chart k-means

df['locs_no'] = df['from_locations']
df.head()

all_cities = ['Chico, CA', 'California, USA', 'Sacramento, CA', 'Reno, NV',
       'Rocklin, CA', 'Folsom, CA', 'Oroville, CA', 'Florin, CA',
       'Nevada City, CA', 'Yountville, CA', 'Redding, CA', 'Linda, CA',
       'Magalia, CA', 'Paradise, CA', 'nolocationfound', 'Durham, CA',
       'Napa, CA', 'Glen Ellen, CA', 'Oroville East, CA',
       'South Oroville, CA', 'Thermalito, CA', 'Gridley, CA']

butte_cty = ['Chico, CA',  'Oroville, CA', 'Magalia, CA', 'Paradise, CA',  
             'Durham, CA', 'Oroville East, CA', 'South Oroville, CA', 
             'Thermalito, CA', 'Gridley, CA']

not_butte = ['California, USA', 'Sacramento, CA', 'Reno, NV','Rocklin, CA', 
             'Folsom, CA', 'Florin, CA','Nevada City, CA', 'Yountville, CA', 
             'Redding, CA', 'Linda, CA', 'nolocationfound', 'Napa, CA', 
             'Glen Ellen, CA',]

df['locs_no'] = df['locs_no'].replace(['Chico, CA',  'Oroville, CA', 'Magalia, CA', 'Paradise, CA',  
             'Durham, CA', 'Oroville East, CA', 'South Oroville, CA', 
             'Thermalito, CA', 'Gridley, CA'], 1)

df['locs_no'] = df['locs_no'].replace(['California, USA', 'Sacramento, CA', 'Reno, NV','Rocklin, CA', 
             'Folsom, CA', 'Florin, CA','Nevada City, CA', 'Yountville, CA', 
             'Redding, CA', 'Linda, CA', 'nolocationfound', 'Napa, CA', 
             'Glen Ellen, CA',], 0)

df.head()

df_cluster = df[['key_score', 'locs_no']]

df['timestamp'].sort_values()

import chart_studio.plotly as py 
import plotly.graph_objs as go 
import plotly.express as px

# Simple line graph of apple high prices
fig = px.line(df, # data
             x = 'timestamp', # date for x-axis
             y = 'key_score' ) # plot close price
fig.show()
